v3: Refactor testing helpers (#231)
* Move `with`-style testing helpers to a separate file

* Move Equatable and Logger extensions to separate files
